bird_demo/pyproject.toml
[tool.poetry]
name = "bird_demo"
version = "0.1.0"
description = "A project that uses birds to demonstrate how metacontext is used"
authors = ["raymondpeterdorn <rpd346@gmail.com>"]
readme = "README.md"
packages = [{include = "bird_demo"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = ">=1.0.0"
geopandas = ">=1.1.1,<2.0.0"
scipy = ">=1.16.2,<2.0.0"
scikit-learn = ">=1.7.2,<2.0.0"
numpy = ">=2.3.3,<3.0.0"
pydantic = ">=2.11.9,<3.0.0"
pillow = ">=11.3.0,<12.0.0"
metacontext = {path = "../metacontext", develop = true}

[tool.poetry.group.dev.dependencies]
mypy = ">=1.17.1"
pre-commit = ">=4.3.0"
pytest = ">=8.4.2"
pytest-cov = ">=4.0"
ruff = ">=0.12.12"
types-PyYAML = ">=6.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
